Move files around and rename to match convention
